requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
reportlab==4.0.7
gunicorn==21.2.0
